+++ src/kredivo/models.py
@@ -1,3 +1,43 @@
+from enum import Enum
+
+__all__ = ['KredivoStatus', 'KredivoTransactionStatus', 'KredivoFraudStatus', 'SerializableMixin',
+           'TransactionDetail', 'ItemDetail', 'CustomerDetail', 'Address', 'KredivoCheckoutEntity',
+           'KredivoTransactionResponse', 'CancelPurchaseResponse', 'KredivoPaymentStatusResponse']
+
+
+# Kredivo Status
+class KredivoStatus:
+    OK = "OK"
+    ERROR = "ERROR"
+
+
+class KredivoTransactionStatus:
+    """
+    Transaction status of a transaction, the values is:
+    CAPTURE: Transaction has been captured
+    SETTLEMENT: Transaction has been settled
+    PENDING: Transaction has not been completed
+    DENY: Transaction has been denied
+    CANCEL: Transaction has been cancelled
+    EXPIRE: Transaction has been expired
+    """
+    PENDING = "pending"
+    SETTLEMENT = "settlement"
+    DENY = "deny"
+    CANCEL = "cancel"
+    EXPIRE = "expire"
+
+
+class KredivoFraudStatus:
+    """
+    Detection result by Fraud Detection System (FDS), the values is:
+    ACCEPT: Approved by FDS
+    DENY: Denied by FDS. Transaction automatically failed
+    """
+    ACCEPT = "accept"
+    DENY = "deny"
+
+
 class SerializableMixin(object):
     """
     An instance that can return a dictionary representation of it's
@@ -65,7 +105,12 @@ class KredivoTransactionResponse(SerializableMixin):
                  "amount", "payment_type", "transaction_status", "message", "transaction_id"]
 
 
-class CancelPurcaseResponse(SerializableMixin):
+class CancelPurchaseResponse(SerializableMixin):
 
     __fields__ = ["status", "fraud_status", "order_id", "transaction_time", "amount", "payment_type",
                   "transaction_status", "message", "transaction_id"]
+
+
+class KredivoPaymentStatusResponse(SerializableMixin):
+    __fields__ = ["status", "legal_name", "fraud_status", "order_id", "transaction_time",
+                 "amount", "payment_type", "transaction_status", "message", "transaction_id"]
